+++ e-token/src/processor/internal/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod ephemeral_ata;
+pub(crate) mod execute_queued_transfer;
 pub(crate) mod group_receipt;
+pub(crate) mod group_receipt_accounts;
 pub(crate) mod lamports_pda;
 pub(crate) mod shuttle_delegation;
 pub(crate) mod token_vault;
